Narrow PRNG counter from u64 to u32 to match spec
The Elixir reference uses <<seed::binary, ctr::32-big>>, making the
counter explicitly 32-bit. The Rust side used u64 and silently
truncated via `ctr as u32`, meaning counters 0 and 2^32 would
produce identical PRNG blocks. Now the API is u32 throughout,
making the spec constraint explicit and overflow impossible to miss.

Closes PAM-731

+++ src/lib.rs
@@ -68,7 +68,7 @@ fn shuffle(pool: &[String], seed: &[u8; 32]) -> Vec<String> {
         return arr;
     }
 
-    let mut ctr: u64 = 0;
+    let mut ctr: u32 = 0;
 
     for i in (1..m).rev() {
         let (j, next_ctr) = prng::random_integer(seed, ctr, (i + 1) as u64);
